requirements.txt
requests
python-dotenv
fastapi[standard]
uvicorn[standard]